Refactor module structure in the main crate
