--- src/resources/mod.rs
@@ -1,2 +1,0 @@
-pub mod noise;
-pub mod socket;
